requirements.txt
requests
python-dotenv
pyyaml
feedparser
pandas
yfinance